feat: create shared contract library crate (#160)
- Add packages/contracts/shared/ with Cargo.toml and src/lib.rs
- Export OUTCOME_UP=1, OUTCOME_DOWN=2, MAX_FEE_BPS=10_000, MESSAGE_PREFIX
- Export build_message(), validate_call_id(), is_valid_outcome(), is_valid_fee_bps()
- Update call_registry and outcome_manager to depend on backit-shared
- Replace magic numbers and duplicated build_message with shared helpers
- All 70 tests pass
