+++ packages/contracts/shared/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "backit-shared"
+version = "0.0.0"
+edition = "2021"
+publish = false
+
+[lib]
+crate-type = ["lib"]
+doctest = false
+
+[dependencies]
+soroban-sdk = { workspace = true }
